RandomCreateDataManager 클래스 생성하여 작업
